+++ flask_accepts/utils_test.py
@@ -21,37 +21,46 @@ def test_unpack_nested():
     assert result
 
 
-def test__default_model_name():
-    from .utils import _default_model_name
+def test_get_default_model_name():
+    from .utils import get_default_model_name
 
     class TestSchema(Schema):
         pass
 
-    result = _default_model_name(TestSchema)
+    result = get_default_model_name(TestSchema)
 
     expected = "Test"
     assert result == expected
 
 
-def test__default_model_name_works_with_multiple_schema_in_name():
-    from .utils import _default_model_name
+def test_get_default_model_name_works_with_multiple_schema_in_name():
+    from .utils import get_default_model_name
 
     class TestSchemaSchema(Schema):
         pass
 
-    result = _default_model_name(TestSchemaSchema)
+    result = get_default_model_name(TestSchemaSchema)
 
     expected = "TestSchema"
     assert result == expected
 
 
-def test__default_model_name_that_does_not_end_in_schema():
-    from .utils import _default_model_name
+def test_get_default_model_name_that_does_not_end_in_schema():
+    from .utils import get_default_model_name
 
     class SomeOtherName(Schema):
         pass
 
-    result = _default_model_name(SomeOtherName)
+    result = get_default_model_name(SomeOtherName)
 
     expected = "SomeOtherName"
     assert result == expected
+
+
+def test_get_default_model_name_default_names():
+    from .utils import get_default_model_name
+
+    for model_num in range(5):
+        result = get_default_model_name()
+        expected = f"DefaultResponseModel_{model_num}"
+        assert result == expected